Rework watcher to host the dash app and allow for backend emergency updates
In case the backend has crashed it will be restarted by systemd so we do not
need to do so in the watcher.
However we still want an emergency update mechanism for development.
We therefore change the watcher to run on port 3000, host the dash app and
allow for use of zip files as created by the github artifacts.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>